+++ screens/Signup.js
@@ -1,12 +1,10 @@
 import React from 'react';
-import styled from 'styled-components/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { useNavigation } from '@react-navigation/native';
 
 import DetailsInfoPage from '../components/signup/DetailsInfoPage';
 import EmailVerificationPage from '../components/signup/EmailVerificationPage';
 import CertificatePage from '../components/signup/CertificatePage';
-import SignupPage from '../components/signup/SignupPage';
+import StartPage from '../components/signup/StartPage';
 import EmailVerificationPage2 from '../components/signup/EmailVerificationPage2';
 
 const NativeStack = createNativeStackNavigator();
@@ -14,7 +12,7 @@ const NativeStack = createNativeStackNavigator();
 const Signup = () => {
   return (
     <NativeStack.Navigator screenOptions={{ headerShown: false }}>
-      <NativeStack.Screen name="SignupPage" component={SignupPage} />
+      <NativeStack.Screen name="StartPage" component={StartPage} />
       <NativeStack.Screen name="DetailsInfoPage" component={DetailsInfoPage} />
       <NativeStack.Screen
         name="EmailVerificationPage"
